projects/explorer-ai/requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.0
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.4.0
attrs==24.1.0
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.7
distro==1.9.0
fastcore==1.6.1
fastlite==0.0.7
frozenlist==1.4.1
groq==0.9.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.18
langchain-community==0.3.17
langchain-core==0.3.49
langchain-nvidia-ai-endpoints==0.3.9
langchain-text-splitters==0.3.6
langgraph==0.3.22
langgraph-checkpoint==2.0.23
langgraph-checkpoint-sqlite==2.0.6
langgraph-prebuilt==0.1.7
langgraph-sdk==0.1.60
langsmith==0.3.22
marshmallow==3.21.3
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
oauthlib==3.2.2
openai==1.38.0
orjson==3.10.6
packaging==24.1
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-fasthtml==0.2.1
python-multipart==0.0.9
PyYAML==6.0.1
regex==2024.7.24
requests==2.32.3
shortuuid==1.0.13
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
sqlite_minutils==3.36.0.post4
starlette==0.38.2
tenacity==8.5.0
tiktoken==0.7.0
tqdm==4.66.4
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.4
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
yarl==1.9.4
